refactor encounter list view
