+++ client/packages/system/src/Encounter/index.ts
@@ -1,3 +1,4 @@
 export * from './ListView';
 export * from './Service';
 export * from './api';
+export * from './utils';
